Changed sampling rate from 10s to 3s.
Some of the pools are sending new jobs every ~20s.  With a 10s sampling rate, we were getting the low hashrate due to job switch and then a good sample.  Averaged together, it looked like a hashrate drop and was causing excessive miner restarts.  No restarts so far after changing to 3s.